refactor: point nsfollow gzip consumers at shared GzipRequestInterceptor (Phase E)
